clean BalanceClient base response typed HttpClient, custom exceptions and fix. Endpoints OK.

+++ src/ECommerce.Application/Abstractions/IBalanceClient.cs
@@ -5,6 +5,6 @@ namespace ECommerce.Application.Abstractions;
 public interface IBalanceClient
 {
     Task<IReadOnlyList<ProductInfo>> GetProductsAsync(CancellationToken cancellationToken);
-    Task<PreorderInfo> PreorderAsync(decimal amount, string orderId, CancellationToken cancellationToken);
+    Task<PreorderInfo> PreorderAsync(int amount, string orderId, CancellationToken cancellationToken);
     Task<CompletionInfo> CompleteAsync(string orderId, CancellationToken cancellationToken);
 }
